v23 - pridana divka a bezdomovec do prislusnych prostoru

+++ src/Adventura/TextoveRozhrani.cpp
@@ -20,17 +20,17 @@ void TextoveRozhrani::hraj(void)
 		// <help box>
 		{
 			{
-				COORD c = { 0, hra.wnd_h - 2 };
+				COORD c = { 0, (short)hra.wnd_h - 2 };
 				SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), c);
 				std::cout << '+' << std::string(hra.wnd_w - 2, '-') << '+';
 			}
 			{
-				COORD c = { 0, hra.wnd_h - 1 };
+				COORD c = { 0, (short)hra.wnd_h - 1 };
 				SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), c);
 				hra.napoveda();
 			}
 			{
-				COORD c = { 0, hra.wnd_h };
+				COORD c = { 0, (short)hra.wnd_h };
 				SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), c);
 				std::cout << '+' << std::string(hra.wnd_w-2, '-') << '+';
 			}
@@ -41,18 +41,18 @@ void TextoveRozhrani::hraj(void)
 		// <command box>
 		{
 			{
-				COORD c = { 0, hra.wnd_h - 5 };
+				COORD c = { 0, (short)hra.wnd_h - 5 };
 				SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), c);
 				std::cout << std::string(hra.wnd_w, '=');
 				std::cout << '\n';
 			}
 			{
-				COORD c = { 0, hra.wnd_h - 3 };
+				COORD c = { 0, (short)hra.wnd_h - 3 };
 				SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), c);
 				std::cout << std::string(hra.wnd_w, '=');
 			}
 			{
-				COORD c = { 0, hra.wnd_h - 4 };
+				COORD c = { 0, (short)hra.wnd_h - 4 };
 				SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), c);
 				std::cout << ">>> ";
 				std::getline(std::cin, prikaz);
